add test cases for HBaseModel.
