+++ s2core/src/main/scala/com/daumkakao/s2graph/core/models/HColumnMeta.scala
@@ -1,29 +1,61 @@
-//package com.daumkakao.s2graph.core.models
-//
-///**
-// * Created by shon on 5/12/15.
-// */
-//case class HColumnMeta(tableName: String = "HColumnMeta", kvs: Map[String, String]) extends HBaseModel(tableName, kvs) {
-//  import HBaseModel._
-//  logicalTableName = "HColumnMeta"
-//  val pk = KeyVals(Seq("columnId", "name"), Seq(columnId, name))
-//  val pkVal = KeyVals(Seq("seq"), Seq(seq))
-//  val idxByColumnIdSeq = KeyVals(Seq("columnId", "seq"), Seq(columnId, seq))
-//  val idxValByColumnIdSeq = KeyVals(Seq("name"), Seq(name))
-//
-//  def insert(zkQuorum: String): Boolean = {
-//    super.insert(zkQuorum)(id, pk, pkVal) &&
-//    super.insert(zkQuorum)(id, idxByColumnIdSeq, idxValByColumnIdSeq)
-//  }
-//  override def parse(rowKey: String, qualifier: String, value: String): HColumnMeta = {
-//
-//    val components = rowKey.split(DELIMITER)
-//    try {
-//
-//      val idxKVs = KeyVals(components.dropRight(1).mkString(DELIMITER)).toKVMap()
-//      val metaKVs = KeyVals(value).toKVMap()
-//      val kvs = idxKVs ++ metaKVs
-//      HColumnMeta(kvs("id").toLong, kvs("columnId").toInt, kvs("name"), kvs("seq").toByte)
-//    }
-//  }
-//}
+package com.daumkakao.s2graph.core.models
+
+import com.daumkakao.s2graph.core.models.HBaseModel.{VAL, KEY}
+
+/**
+ * Created by shon on 5/15/15.
+ */
+
+object HColumnMeta {
+  val timeStampSeq = 0.toByte
+  val countSeq = -1.toByte
+  val lastModifiedAtColumnSeq = 0.toByte
+  val lastModifiedAtColumn = HColumnMeta(Map("id" -> 0, "columnId" -> 0,
+    "name" -> "lastModifiedAt", "seq" -> lastModifiedAtColumnSeq))
+  val maxValue = Byte.MaxValue
+
+
+  def findById(id: Int, useCache: Boolean = true): HColumnMeta = {
+    HBaseModel.find("HColumnMeta", useCache)(Seq(("id" -> id))).get.asInstanceOf[HColumnMeta]
+  }
+  def findAllByColumn(columnId: Int, useCache: Boolean = true) = {
+    HBaseModel.findsMatch("HColumnMeta", useCache)(Seq(("columnId" -> columnId))).map(x => x.asInstanceOf[HColumnMeta])
+  }
+  def findByName(columnId: Int, name: String, useCache: Boolean = true) = {
+    HBaseModel.find("HColumnMeta", useCache)(Seq(("columnId" -> columnId), ("name" -> name))).map(x => x.asInstanceOf[HColumnMeta])
+  }
+  def findByIdAndSeq(columnId: Int, seq: Byte, useCache: Boolean = true) = {
+    HBaseModel.find("HColumnMeta", useCache)(Seq(("columnId" -> columnId), ("seq" -> seq))).map(x => x.asInstanceOf[HColumnMeta])
+  }
+  def findOrInsert(columnId: Int, name: String): HColumnMeta = {
+    findByName(columnId, name, useCache = false) match {
+      case Some(s) => s
+      case None =>
+        val id = HBaseModel.getAndIncrSeq("HColumnMeta")
+        val allMetas = findAllByColumn(columnId, useCache = false)
+        val seq = (allMetas.length + 1).toByte
+        val model = HColumnMeta(Map("id" -> id, "columnId" -> columnId, "name" -> name, "seq" -> seq))
+        model.create
+        model
+    }
+  }
+}
+
+case class HColumnMeta(kvsParam: Map[KEY, VAL]) extends HBaseModel("HColumnMeta", kvsParam) {
+  override val columns = Seq("id", "columnId", "name", "seq")
+
+  val pk = Seq(("id", kvs("id")))
+  val idxColumnIdName = Seq(("columnId", kvs("columnId")), ("name", kvs("name")))
+  val idxColumnIdSeq = Seq(("columnId", kvs("columnId")), ("seq", kvs("seq")))
+
+  override val idxKVsList = List(pk, idxColumnIdName, idxColumnIdSeq)
+  validate(columns)
+
+  val id = Some(kvs("id").toString.toInt)
+  val columnId = kvs("columnId").toString.toInt
+  val name = kvs("name").toString
+  val seq = kvs("seq").toString.toByte
+
+}
+
+
