+++ s2core/src/main/scala/com/daumkakao/s2graph/core/models/HServiceColumn.scala
@@ -1,32 +1,52 @@
-//package com.daumkakao.s2graph.core.models
-//
-//import com.daumkakao.s2graph.core.models.HBaseModel._
-//
-///**
-// * Created by shon on 5/12/15.
-// */
-////ALTER TABLE service_columns add FOREIGN KEY(service_id) REFERENCES services(id) ON DELETE CASCADE;
-//
-//case class HServiceColumn(id: Long, serviceId: Int, columnName: String, columnType: String) extends HBaseModel {
-//  logicalTableName = "HServiceColumn"
-//  val pk = KeyVals(Seq("id"), Seq(id))
-//  val pkVal = KeyVals(Seq("serviceId", "columnName", "columnType"), Seq(serviceId, columnName, columnType))
-//
-//  val idxByServiceIdColumnName = KeyVals(Seq("serviceId", "columnName"), Seq(serviceId, columnName))
-//  val idxValByServiceIdColumnName = KeyVals(Seq("columnType"), Seq(columnType))
-//
-//  def insert(zkQuorum: String) =
-//    super.insert(zkQuorum)(id, pk, pkVal) &&
-//    super.insert(zkQuorum)(id, idxByServiceIdColumnName, idxValByServiceIdColumnName)
-//
-//  override def parse(rowKey: String, qualifier: String, value: String): HServiceColumn = {
-//    val components = rowKey.split(DELIMITER)
-//    println(rowKey, qualifier, value)
-//    try {
-//      val idxKVs = KeyVals(components.dropRight(1).mkString(DELIMITER)).toKVMap()
-//      val metaKVs = KeyVals(value).toKVMap()
-//      val kvs = idxKVs ++ metaKVs
-//      HServiceColumn(kvs("id").toLong, kvs("serviceId").toInt, kvs("columnName"), kvs("columnType"))
-//    }
-//  }
-//}
+package com.daumkakao.s2graph.core.models
+
+import HBaseModel._
+import play.api.libs.json.Json
+
+/**
+ * Created by shon on 5/15/15.
+ */
+
+object HServiceColumn {
+  def findById(id: Int, useCache: Boolean = true): HServiceColumn = {
+    HBaseModel.find("HServiceColumn", useCache)(Seq(("id" -> id))).get.asInstanceOf[HServiceColumn]
+  }
+  def find(serviceId: Int, columnName: String, useCache: Boolean = true): Option[HServiceColumn] = {
+    HBaseModel.find("HServiceColumn", useCache)(Seq("serviceId" -> serviceId, "columnName" -> columnName))
+      .map { x => x.asInstanceOf[HServiceColumn]}
+  }
+  def findsByServiceId(serviceId: Int, useCache: Boolean = true): List[HServiceColumn] = {
+    HBaseModel.findsMatch("HServiceColumn", useCache)(Seq("serviceId" -> serviceId)).map(x => x.asInstanceOf[HServiceColumn])
+  }
+  def findOrInsert(serviceId: Int, columnName: String, columnType: Option[String]): HServiceColumn = {
+    find(serviceId, columnName, useCache = false) match {
+      case Some(s) => s
+      case None =>
+        val id = HBaseModel.getAndIncrSeq("HServiceColumn")
+        val model = new HServiceColumn(Map("id" -> id, "serviceId" -> serviceId, "columnName" -> columnName,
+          "columnType" -> columnType.getOrElse("string")))
+        model.create
+        model
+    }
+  }
+}
+case class HServiceColumn(kvsParam: Map[KEY, VAL]) extends HBaseModel("HServiceColumn", kvsParam) {
+  override val columns = Seq("id", "serviceId", "columnName", "columnType")
+  val pk = Seq(("id", kvs("id")))
+  val idxServiceIdColumnName = Seq(("serviceId", kvs("serviceId")), ("columnName", kvs("columnName")))
+  override val idxKVsList = List(pk, idxServiceIdColumnName)
+  override val referencedBysList = List(Seq(("HColumnMeta", "columnId", "id")))
+  validate(columns)
+
+  val id = Some(kvs("id").toString.toInt)
+  val serviceId = kvs("serviceId").toString.toInt
+  val columnName = kvs("columnName").toString
+  val columnType = kvs("columnType").toString
+
+
+  lazy val service = HService.findById(serviceId)
+  lazy val metas = HColumnMeta.findAllByColumn(id.get)
+  lazy val metaNamesMap = (HColumnMeta.lastModifiedAtColumn :: metas).map(x => (x.seq, x.name)) toMap
+  lazy val toJson = Json.obj("serviceName" -> service.serviceName, "columnName" -> columnName, "columnType" -> columnType)
+
+}
