+++ s2core/src/test/scala/com/daumkakao/s2graph/core/ManagementTest.scala
@@ -1,6 +1,6 @@
 package com.daumkakao.s2graph.core
 import com.daumkakao.s2graph.core.Graph
-import com.daumkakao.s2graph.core.models.HBaseModel
+import com.daumkakao.s2graph.core.models.{HLabel, HService, HServiceColumn, HBaseModel}
 import com.typesafe.config.ConfigFactory
 import org.scalatest.{FunSuite, Matchers}
 import play.api.libs.json.{JsString, JsBoolean, JsNumber, Json}
@@ -26,23 +26,98 @@ class ManagementTest extends FunSuite with Matchers {
   val config = ConfigFactory.parseString(s"hbase.zookeeper.quorum=$zkQuorum")
   Graph(config)(ExecutionContext.Implicits.global)
   HBaseModel(zkQuorum)
+  val current = System.currentTimeMillis()
+  val serviceNames = (0 until 2).map { i => s"$serviceName-${current + i}" }
+  val labelNames = (0 until 2).map { i => s"$labelName-${current + i}" }
 
+//  def runTC[T <: HBaseModel](prevSeq: Long, prevSize: Int, prefix: String)(testSize: Int)(createF: String => T)(fetchF: String => Option[T])(deleteF: String => Boolean) = {
+//    var lastSeq = prevSeq
+//    val createds = collection.mutable.Map.empty[String, T]
+//    val names = (0 until testSize) map { i => s"$prefix-${current + i}"}
+//
+//    val rets = for {
+//      name <- names
+//    } yield {
+//        val created = createF(name)
+//        val testSeq = created.id.get > lastSeq
+//        lastSeq = created.id.get
+//        createds += (name -> created)
+//        val fetched = fetchF(name)
+//        fetched.isDefined && created == fetched.get && testSeq
+//      }
+//
+//    val deletes = for {
+//      name <- names
+//    } yield {
+//        deleteF(name)
+//      }
+//
+//    (rets ++ deletes).forall(_)
+//  }
   test("test create service") {
-    Management.deleteService(serviceName)
-    Management.findService(serviceName) == None
-    val service = Management.createService(serviceName, zkQuorum, hTableName.get, preSplitSize, hTableTTL)
-    val other = Management.findService(service.serviceName)
-    other.isDefined && service == other.get
+
+    var prevSeq = Management.getSequence("HService")
+    val prevSize = HService.findAllServices().size
+    val createds = collection.mutable.Map.empty[String, HService]
+
+    val rets = for {
+      serviceName <- serviceNames
+    } yield {
+      val service = Management.createService(serviceName, zkQuorum, hTableName.get, preSplitSize, hTableTTL)
+      val testSeq = service.id.get > prevSeq
+      prevSeq = service.id.get
+      createds += (service.serviceName -> service)
+      val other = Management.findService(service.serviceName)
+      other.isDefined && service == other.get && testSeq
+    }
+
+    val deletes = for {
+      serviceName <- serviceNames
+    } yield {
+      Management.deleteService(serviceName)
+    }
+    (rets ++ deletes).forall(_)
+
+    HService.findAllServices().size == prevSize
   }
   test("test create label") {
-    Management.deleteLabel(labelName)
-    Management.findLabel(labelName) == None
-    val label = Management.createLabel(labelName, serviceName, columnName, columnType,
-      serviceName, columnName, columnType,
-      true, serviceName, indexProps, props,
-      consistencyLevel, hTableName, hTableTTL
-    )
-    val other = Management.findLabel(labelName)
-    other.isDefined && label == other.get
+    val service = Management.createService(serviceName, zkQuorum, hTableName.get, preSplitSize, hTableTTL)
+    var prevSeq = Management.getSequence("HLabel")
+    val prevSize = HLabel.findAll(useCache = false)
+    val createds = collection.mutable.Map.empty[String, HLabel]
+
+    val rets = for {
+      lName <- labelNames
+    } yield {
+      val label = Management.createLabel(lName, serviceName, columnName, columnType,
+        serviceName, columnName, columnType,
+        true, serviceName, indexProps, props,
+        consistencyLevel, hTableName, hTableTTL
+      )
+      val testSeq = label.id.get > prevSeq
+      prevSeq = label.id.get
+
+      createds += (label.label -> label)
+      val other = Management.findLabel(label.label)
+      other.isDefined && label == other.get && testSeq
+    }
+    println(HLabel.findAll(useCache = false))
+    val deletes = for {
+      lName <- labelNames
+    } yield {
+        Management.deleteLabel(lName)
+      }
+    (rets ++ deletes).forall(_)
+    HLabel.findAll(useCache = false).size == prevSize
   }
+
+//  test("test delete label") {
+//    Management.deleteLabel(s"$labelName-0")
+//    Management.findLabel(s"$labelName-0") == None
+//  }
+//  test("test delete service") {
+//    Management.deleteService(serviceName)
+//    Management.findService(serviceName) == None
+//  }
+
 }
